Add footprint for BetLux BL-FL7680RGB "Piranha" RGB LED.
https://cdn-shop.adafruit.com/datasheets/BL-FL7680RGB.pdf

The datasheet does not appear to be very good.  Even though this is a
datasheet for an RGB LED, it labels two of the fours pins as "C" and
the other two pins as "A".  So, I had to determine the pinout
experimentally.

The pin in the corner with the chamfer is blue.  Since the datasheet
does not give any pin numbering scheme, I figured the pin with the
chamfer ought to be pin 1.  From there, there are two reasonable
pinouts: either clockwise (BARG) or counterclockwise (BGRA).  I
arbitrarily chose counterclockwise.